requirements.txt
django-bootstrap-toolkit
django-less
fuzzywuzzy
PIL
South
